requirements.txt
alembic==1.7.7
bidict==0.22.0
cachelib==0.6.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
simple-websocket==0.5.2
SQLAlchemy==1.4.36
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
